requirements.txt
tensorflow>=2.10.0
numpy>=1.21.0
scikit-learn>=1.0.0
matplotlib>=3.5.0
seaborn>=0.11.0
flask>=2.0.0
pillow>=8.0.0
quickdraw>=0.1.0
